Re-add support for reading cg entity data
To allow reading indices created with pre 4.0.0 versions of
metafacture-search, support for reading cg entity data is re-added. A
heuristic is applied to decide whether the serialized data on the index
is in Formeta or CG Entity format. If the serialised record contains
field and subfield markers (characters 0x1e and 0x1f) it is assumed that
it is encoded with CG entity otherwise Formeta encoding is assumed.